split loadActivation and activation code, add tons of dev logging
